Sprint23_add_icon_and_info_for_playlist

+++ app/src/main/java/com/practicum/playlistmaker/media/data/dto/PlaylistDto.kt
@@ -5,6 +5,6 @@ data class PlaylistDto(
     val playlistName: String,
     val playlistDescription: String?,
     val imagePath: String?,
-    val listIdTracks: List<Long>,
+    val listIdTracks: List<Int>,
     val numberOfTracks: Int
 )
